feat: 优化因子库管理和策略管理
