+++ backend/app/db/database.py
@@ -56,7 +56,8 @@ def init_db():
     try:
         # 直接导入所有模型以确保它们被注册到Base.metadata
         from app.db.models.stock import Stock, StockDaily, MarketIndex
-        from app.db.models.factor import Factor, FactorFormulaHistory, Strategy, StrategyExecution, SelectionResult, FactorValue
+        from app.db.models.factor import Factor, FactorFormulaHistory, FactorValue
+        from app.db.models.strategy import Strategy, StrategyExecution, SelectionResult
         from app.db.models.watchlist import Watchlist, WatchlistGroup, WatchlistGroupMember
         
         # 创建所有表
