+++ backend/app/db/models/factor.py
@@ -71,104 +71,6 @@ class FactorFormulaHistory(Base):
         return f"<FactorFormulaHistory(id={self.id}, factor_id='{self.factor_id}')>"
 
 
-class Strategy(Base):
-    """策略模型"""
-    __tablename__ = "strategies"
-    
-    id = Column(Integer, primary_key=True, index=True)
-    name = Column(String(200), nullable=False, comment="策略名称")
-    description = Column(Text, comment="策略描述")
-    
-    # 因子配置
-    factor_ids = Column(JSON, comment="使用的因子ID列表")
-    factor_weights = Column(JSON, comment="因子权重配置")
-    
-    # 筛选条件
-    max_results = Column(Integer, default=50, comment="最大结果数量")
-    min_market_cap = Column(Float, comment="最小市值")
-    max_market_cap = Column(Float, comment="最大市值")
-    exclude_st = Column(Boolean, default=True, comment="排除ST股票")
-    exclude_new_stock = Column(Boolean, default=True, comment="排除新股")
-    min_turnover = Column(Float, comment="最小换手率")
-    
-    # 状态字段
-    is_active = Column(Boolean, default=True, comment="是否启用")
-    
-    # 统计字段
-    execution_count = Column(Integer, default=0, comment="执行次数")
-    last_executed_at = Column(DateTime, comment="最后执行时间")
-    avg_execution_time = Column(Float, comment="平均执行时间")
-    
-    # 时间字段
-    created_at = Column(DateTime, server_default=func.now(), comment="创建时间")
-    updated_at = Column(DateTime, server_default=func.now(), onupdate=func.now(), comment="更新时间")
-    
-    def __repr__(self):
-        return f"<Strategy(id={self.id}, name='{self.name}')>"
-
-
-class StrategyExecution(Base):
-    """策略执行记录"""
-    __tablename__ = "strategy_executions"
-    
-    id = Column(Integer, primary_key=True, index=True)
-    strategy_id = Column(Integer, ForeignKey("strategies.id"), nullable=False, comment="策略ID")
-    
-    # 执行信息
-    execution_date = Column(DateTime, nullable=False, comment="执行日期")
-    execution_time = Column(Float, comment="执行耗时(秒)")
-    total_stocks = Column(Integer, comment="总股票数")
-    result_count = Column(Integer, comment="结果数量")
-    
-    # 执行状态
-    status = Column(String(20), default="completed", comment="执行状态")
-    error_message = Column(Text, comment="错误信息")
-    
-    # 统计信息
-    statistics = Column(JSON, comment="执行统计信息")
-    
-    # 时间字段
-    created_at = Column(DateTime, server_default=func.now(), comment="创建时间")
-    
-    # 关联关系
-    strategy = relationship("Strategy", backref="executions")
-    
-    def __repr__(self):
-        return f"<StrategyExecution(id={self.id}, strategy_id={self.strategy_id})>"
-
-
-class SelectionResult(Base):
-    """选股结果"""
-    __tablename__ = "selection_results"
-    
-    id = Column(Integer, primary_key=True, index=True)
-    execution_id = Column(Integer, ForeignKey("strategy_executions.id"), nullable=False, comment="执行ID")
-    
-    # 股票信息
-    symbol = Column(String(20), nullable=False, comment="股票代码")
-    name = Column(String(100), comment="股票名称")
-    
-    # 评分信息
-    total_score = Column(Float, comment="总评分")
-    rank = Column(Integer, comment="排名")
-    factor_scores = Column(JSON, comment="各因子得分")
-    
-    # 股票属性
-    price = Column(Float, comment="当前价格")
-    market_cap = Column(Float, comment="市值")
-    pe_ratio = Column(Float, comment="市盈率")
-    pb_ratio = Column(Float, comment="市净率")
-    
-    # 时间字段
-    created_at = Column(DateTime, server_default=func.now(), comment="创建时间")
-    
-    # 关联关系
-    execution = relationship("StrategyExecution", backref="results")
-    
-    def __repr__(self):
-        return f"<SelectionResult(id={self.id}, symbol='{self.symbol}', rank={self.rank})>"
-
-
 class FactorValue(Base):
     """因子值记录"""
     __tablename__ = "factor_values"
