+++ frontend/src/components/FactorCreateModalImproved.tsx
@@ -1,6 +1,11 @@
 import React, { useState } from 'react';
 import Editor from '@monaco-editor/react';
-import { XMarkIcon, SparklesIcon, CheckCircleIcon, ExclamationTriangleIcon } from '@heroicons/react/24/outline';
+import {
+  XMarkIcon,
+  SparklesIcon,
+  CheckCircleIcon,
+  ExclamationTriangleIcon,
+} from '@heroicons/react/24/outline';
 import { factorApi, CustomFactorCreateRequest } from '../services/api';
 import FieldSelector from './FieldSelector';
 
@@ -13,7 +18,7 @@ interface FactorCreateModalImprovedProps {
 const FactorCreateModalImproved: React.FC<FactorCreateModalImprovedProps> = ({
   isOpen,
   onClose,
-  onCreated
+  onCreated,
 }) => {
   const [createFactorForm, setCreateFactorForm] = useState({
     factor_id: '',
@@ -24,7 +29,7 @@ const FactorCreateModalImproved: React.FC<FactorCreateModalImprovedProps> = ({
     formula: '',
     input_fields: ['close'],
     default_parameters: {},
-    calculation_method: 'formula'
+    calculation_method: 'formula',
   });
   const [isCreating, setIsCreating] = useState(false);
   const [currentStep, setCurrentStep] = useState(1);
@@ -33,40 +38,64 @@ const FactorCreateModalImproved: React.FC<FactorCreateModalImprovedProps> = ({
 
   // 因子分类选项
   const categoryOptions = [
-    { value: 'custom', label: '自定义因子', description: '用户自定义的量化因子' },
-    { value: 'trend', label: '趋势类因子', description: '捕捉价格趋势和动量的因子' },
-    { value: 'momentum', label: '动量类因子', description: '基于价格动量的因子' },
-    { value: 'volume', label: '成交量类因子', description: '基于成交量分析的因子' },
-    { value: 'volatility', label: '波动率类因子', description: '衡量价格波动的因子' },
-    { value: 'value', label: '估值类因子', description: '基于基本面估值的因子' }
+    {
+      value: 'custom',
+      label: '自定义因子',
+      description: '用户自定义的量化因子',
+    },
+    {
+      value: 'trend',
+      label: '趋势类因子',
+      description: '捕捉价格趋势和动量的因子',
+    },
+    {
+      value: 'momentum',
+      label: '动量类因子',
+      description: '基于价格动量的因子',
+    },
+    {
+      value: 'volume',
+      label: '成交量类因子',
+      description: '基于成交量分析的因子',
+    },
+    {
+      value: 'volatility',
+      label: '波动率类因子',
+      description: '衡量价格波动的因子',
+    },
+    {
+      value: 'value',
+      label: '估值类因子',
+      description: '基于基本面估值的因子',
+    },
   ];
 
   // 验证表单
   const validateForm = (): string[] => {
     const errors: string[] = [];
-    
+
     if (!createFactorForm.factor_id.trim()) {
       errors.push('请填写因子ID');
     } else if (!/^[a-zA-Z][a-zA-Z0-9_]*$/.test(createFactorForm.factor_id)) {
       errors.push('因子ID只能包含字母、数字和下划线，且必须以字母开头');
     }
-    
+
     if (!createFactorForm.name.trim()) {
       errors.push('请填写因子名称');
     }
-    
+
     if (!createFactorForm.display_name.trim()) {
       errors.push('请填写显示名称');
     }
-    
+
     if (!createFactorForm.formula.trim()) {
       errors.push('请编写因子代码');
     }
-    
+
     if (createFactorForm.input_fields.length === 0) {
       errors.push('请至少选择一个输入字段');
     }
-    
+
     return errors;
   };
 
@@ -77,11 +106,11 @@ const FactorCreateModalImproved: React.FC<FactorCreateModalImprovedProps> = ({
       setValidationErrors(errors);
       return;
     }
-    
+
     try {
       setIsCreating(true);
       setValidationErrors([]);
-      
+
       const factorData: CustomFactorCreateRequest = {
         factor_id: createFactorForm.factor_id,
         name: createFactorForm.name,
@@ -91,11 +120,11 @@ const FactorCreateModalImproved: React.FC<FactorCreateModalImprovedProps> = ({
         formula: createFactorForm.formula,
         input_fields: createFactorForm.input_fields,
         default_parameters: createFactorForm.default_parameters,
-        calculation_method: createFactorForm.calculation_method
+        calculation_method: createFactorForm.calculation_method,
       };
-      
+
       await factorApi.createFactor(factorData);
-      
+
       // 显示成功动画
       setShowSuccessAnimation(true);
       setTimeout(() => {
@@ -104,7 +133,6 @@ const FactorCreateModalImproved: React.FC<FactorCreateModalImprovedProps> = ({
         onCreated();
         handleResetForm();
       }, 1500);
-      
     } catch (error) {
       console.error('创建因子失败:', error);
       setValidationErrors(['创建因子失败，请检查网络连接或稍后重试']);
@@ -124,7 +152,7 @@ const FactorCreateModalImproved: React.FC<FactorCreateModalImprovedProps> = ({
       formula: '',
       input_fields: ['close'],
       default_parameters: {},
-      calculation_method: 'formula'
+      calculation_method: 'formula',
     });
     setCurrentStep(1);
     setValidationErrors([]);
@@ -133,12 +161,15 @@ const FactorCreateModalImproved: React.FC<FactorCreateModalImprovedProps> = ({
   // 生成默认代码模板
   const generateDefaultCode = () => {
     const factorId = createFactorForm.factor_id || 'custom_factor';
-    const inputFields = createFactorForm.input_fields.length > 0 
-      ? createFactorForm.input_fields 
-      : ['close'];
-    
-    const dataAccess = inputFields.map(field => `    $${field} = data['$${field}']`).join('\n');
-    
+    const inputFields =
+      createFactorForm.input_fields.length > 0
+        ? createFactorForm.input_fields
+        : ['close'];
+
+    const dataAccess = inputFields
+      .map((field) => `    $${field} = data['$${field}']`)
+      .join('\n');
+
     return `def calculate_${factorId}(data):
     """
     ${createFactorForm.description || '自定义因子计算函数'}
@@ -147,7 +178,9 @@ const FactorCreateModalImproved: React.FC<FactorCreateModalImprovedProps> = ({
         data: pandas.DataFrame - 包含股票历史数据的DataFrame
         
     可用字段:
-${inputFields.map(field => `        - $${field}: ${getFieldDescription(field)}`).join('\n')}
+${inputFields
+  .map((field) => `        - $${field}: ${getFieldDescription(field)}`)
+  .join('\n')}
     
     返回:
         pandas.Series - 计算得到的因子值
@@ -174,15 +207,15 @@ ${dataAccess}
   // 获取字段描述
   const getFieldDescription = (fieldName: string): string => {
     const descriptions: Record<string, string> = {
-      'open': '开盘价',
-      'high': '最高价',
-      'low': '最低价',
-      'close': '收盘价',
-      'volume': '成交量',
-      'amount': '成交额',
-      'pre_close': '前收盘价',
-      'change': '涨跌额',
-      'pct_change': '涨跌幅'
+      open: '开盘价',
+      high: '最高价',
+      low: '最低价',
+      close: '收盘价',
+      volume: '成交量',
+      amount: '成交额',
+      pre_close: '前收盘价',
+      change: '涨跌额',
+      pct_change: '涨跌幅',
     };
     return descriptions[fieldName] || fieldName;
   };
@@ -193,21 +226,26 @@ ${dataAccess}
       <div className="flex items-center space-x-2">
         {[1, 2, 3].map((step) => (
           <React.Fragment key={step}>
-            <div className={`
+            <div
+              className={`
               w-8 h-8 rounded-full flex items-center justify-center text-sm font-medium
               transition-all duration-300
-              ${currentStep >= step 
-                ? 'bg-primary text-primary-content' 
-                : 'bg-base-300 text-base-content/60'
+              ${
+                currentStep >= step
+                  ? 'bg-primary text-primary-content'
+                  : 'bg-base-300 text-base-content/60'
               }
-            `}>
+            `}
+            >
               {step}
             </div>
             {step < 3 && (
-              <div className={`
+              <div
+                className={`
                 w-12 h-0.5 transition-all duration-300
                 ${currentStep > step ? 'bg-primary' : 'bg-base-300'}
-              `} />
+              `}
+              />
             )}
           </React.Fragment>
         ))}
@@ -218,14 +256,16 @@ ${dataAccess}
   if (!isOpen) return null;
 
   return (
-    <div className="modal modal-open backdrop-blur-sm">
+    <dialog className="modal modal-open">
       <div className="modal-box max-w-6xl max-h-[90vh] bg-base-100 border border-base-300 shadow-2xl">
         {/* 成功动画覆盖层 */}
         {showSuccessAnimation && (
           <div className="absolute inset-0 bg-base-100/95 flex items-center justify-center z-50 rounded-lg">
             <div className="text-center">
               <CheckCircleIcon className="w-16 h-16 text-success mx-auto mb-4 animate-pulse" />
-              <h3 className="text-xl font-bold text-success mb-2">创建成功！</h3>
+              <h3 className="text-xl font-bold text-success mb-2">
+                创建成功！
+              </h3>
               <p className="text-base-content/70">因子已成功添加到因子库</p>
             </div>
           </div>
@@ -259,7 +299,9 @@ ${dataAccess}
               <h4 className="font-medium">请修正以下问题:</h4>
               <ul className="list-disc list-inside mt-1">
                 {validationErrors.map((error, index) => (
-                  <li key={index} className="text-sm">{error}</li>
+                  <li key={index} className="text-sm">
+                    {error}
+                  </li>
                 ))}
               </ul>
             </div>
@@ -271,25 +313,33 @@ ${dataAccess}
           {currentStep === 1 && (
             <div className="space-y-4">
               <div className="text-center mb-6">
-                <h4 className="text-lg font-semibold text-base-content mb-2">基本信息</h4>
-                <p className="text-base-content/70">设置因子的基本标识和描述信息</p>
+                <h4 className="text-lg font-semibold text-base-content mb-2">
+                  基本信息
+                </h4>
+                <p className="text-base-content/70">
+                  设置因子的基本标识和描述信息
+                </p>
               </div>
 
               <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
                 <div className="form-control">
                   <label className="label">
                     <span className="label-text font-medium">因子ID *</span>
-                    <span className="label-text-alt text-xs">用于代码中的标识符</span>
+                    <span className="label-text-alt text-xs">
+                      用于代码中的标识符
+                    </span>
                   </label>
                   <input
                     type="text"
                     className="input input-bordered bg-base-100 focus:border-primary"
                     placeholder="例如: momentum_factor_001"
                     value={createFactorForm.factor_id}
-                    onChange={(e) => setCreateFactorForm({
-                      ...createFactorForm,
-                      factor_id: e.target.value
-                    })}
+                    onChange={(e) =>
+                      setCreateFactorForm({
+                        ...createFactorForm,
+                        factor_id: e.target.value,
+                      })
+                    }
                   />
                 </div>
 
@@ -303,44 +353,54 @@ ${dataAccess}
                     className="input input-bordered bg-base-100 focus:border-primary"
                     placeholder="例如: Custom Momentum Factor"
                     value={createFactorForm.name}
-                    onChange={(e) => setCreateFactorForm({
-                      ...createFactorForm,
-                      name: e.target.value
-                    })}
+                    onChange={(e) =>
+                      setCreateFactorForm({
+                        ...createFactorForm,
+                        name: e.target.value,
+                      })
+                    }
                   />
                 </div>
 
                 <div className="form-control">
                   <label className="label">
                     <span className="label-text font-medium">显示名称 *</span>
-                    <span className="label-text-alt text-xs">界面显示的中文名称</span>
+                    <span className="label-text-alt text-xs">
+                      界面显示的中文名称
+                    </span>
                   </label>
                   <input
                     type="text"
                     className="input input-bordered bg-base-100 focus:border-primary"
                     placeholder="例如: 自定义动量因子"
                     value={createFactorForm.display_name}
-                    onChange={(e) => setCreateFactorForm({
-                      ...createFactorForm,
-                      display_name: e.target.value
-                    })}
+                    onChange={(e) =>
+                      setCreateFactorForm({
+                        ...createFactorForm,
+                        display_name: e.target.value,
+                      })
+                    }
                   />
                 </div>
 
                 <div className="form-control">
                   <label className="label">
                     <span className="label-text font-medium">因子分类</span>
-                    <span className="label-text-alt text-xs">选择合适的分类</span>
+                    <span className="label-text-alt text-xs">
+                      选择合适的分类
+                    </span>
                   </label>
                   <select
                     className="select select-bordered bg-base-100 focus:border-primary"
                     value={createFactorForm.category}
-                    onChange={(e) => setCreateFactorForm({
-                      ...createFactorForm,
-                      category: e.target.value
-                    })}
+                    onChange={(e) =>
+                      setCreateFactorForm({
+                        ...createFactorForm,
+                        category: e.target.value,
+                      })
+                    }
                   >
-                    {categoryOptions.map(option => (
+                    {categoryOptions.map((option) => (
                       <option key={option.value} value={option.value}>
                         {option.label}
                       </option>
@@ -352,16 +412,20 @@ ${dataAccess}
               <div className="form-control">
                 <label className="label">
                   <span className="label-text font-medium">因子描述</span>
-                  <span className="label-text-alt text-xs">详细说明因子的用途和计算逻辑</span>
+                  <span className="label-text-alt text-xs">
+                    详细说明因子的用途和计算逻辑
+                  </span>
                 </label>
                 <textarea
                   className="textarea textarea-bordered bg-base-100 focus:border-primary h-24 resize-none"
                   placeholder="请描述因子的计算原理、应用场景和预期效果..."
                   value={createFactorForm.description}
-                  onChange={(e) => setCreateFactorForm({
-                    ...createFactorForm,
-                    description: e.target.value
-                  })}
+                  onChange={(e) =>
+                    setCreateFactorForm({
+                      ...createFactorForm,
+                      description: e.target.value,
+                    })
+                  }
                 />
               </div>
 
@@ -369,7 +433,11 @@ ${dataAccess}
                 <button
                   className="btn btn-primary"
                   onClick={() => setCurrentStep(2)}
-                  disabled={!createFactorForm.factor_id || !createFactorForm.name || !createFactorForm.display_name}
+                  disabled={
+                    !createFactorForm.factor_id ||
+                    !createFactorForm.name ||
+                    !createFactorForm.display_name
+                  }
                 >
                   下一步：选择数据字段
                 </button>
@@ -381,21 +449,29 @@ ${dataAccess}
           {currentStep === 2 && (
             <div className="space-y-4">
               <div className="text-center mb-6">
-                <h4 className="text-lg font-semibold text-base-content mb-2">数据字段配置</h4>
-                <p className="text-base-content/70">选择因子计算所需的数据字段</p>
+                <h4 className="text-lg font-semibold text-base-content mb-2">
+                  数据字段配置
+                </h4>
+                <p className="text-base-content/70">
+                  选择因子计算所需的数据字段
+                </p>
               </div>
 
               <div className="form-control">
                 <label className="label">
                   <span className="label-text font-medium">输入字段 *</span>
-                  <span className="label-text-alt text-xs">因子计算时可以使用的数据字段</span>
+                  <span className="label-text-alt text-xs">
+                    因子计算时可以使用的数据字段
+                  </span>
                 </label>
                 <FieldSelector
                   selectedFields={createFactorForm.input_fields}
-                  onChange={(fields) => setCreateFactorForm({
-                    ...createFactorForm,
-                    input_fields: fields
-                  })}
+                  onChange={(fields) =>
+                    setCreateFactorForm({
+                      ...createFactorForm,
+                      input_fields: fields,
+                    })
+                  }
                   placeholder="请选择因子计算需要的数据字段..."
                   showValidation={true}
                 />
@@ -404,12 +480,21 @@ ${dataAccess}
               {/* 字段预览 */}
               {createFactorForm.input_fields.length > 0 && (
                 <div className="bg-base-200/50 rounded-lg p-4">
-                  <h5 className="font-medium text-base-content mb-3">已选择的字段预览</h5>
+                  <h5 className="font-medium text-base-content mb-3">
+                    已选择的字段预览
+                  </h5>
                   <div className="grid grid-cols-2 md:grid-cols-3 gap-3">
-                    {createFactorForm.input_fields.map(field => (
-                      <div key=${field} className="bg-base-100 rounded p-2 text-sm">
-                        <span className="font-medium text-primary">data['${field}']</span>
-                        <div className="text-base-content/70">{getFieldDescription(field)}</div>
+                    {createFactorForm.input_fields.map((field) => (
+                      <div
+                        key={field}
+                        className="bg-base-100 rounded p-2 text-sm"
+                      >
+                        <span className="font-medium text-primary">
+                          data['${field}']
+                        </span>
+                        <div className="text-base-content/70">
+                          {getFieldDescription(field)}
+                        </div>
                       </div>
                     ))}
                   </div>
@@ -438,7 +523,9 @@ ${dataAccess}
           {currentStep === 3 && (
             <div className="space-y-4">
               <div className="text-center mb-6">
-                <h4 className="text-lg font-semibold text-base-content mb-2">编写因子代码</h4>
+                <h4 className="text-lg font-semibold text-base-content mb-2">
+                  编写因子代码
+                </h4>
                 <p className="text-base-content/70">实现因子的计算逻辑</p>
               </div>
 
@@ -450,25 +537,29 @@ ${dataAccess}
                   <button
                     type="button"
                     className="btn btn-sm btn-outline btn-primary"
-                    onClick={() => setCreateFactorForm({
-                      ...createFactorForm,
-                      formula: generateDefaultCode()
-                    })}
+                    onClick={() =>
+                      setCreateFactorForm({
+                        ...createFactorForm,
+                        formula: generateDefaultCode(),
+                      })
+                    }
                   >
                     <SparklesIcon className="w-4 h-4 mr-1" />
                     生成模板
                   </button>
                 </div>
-                
+
                 <div className="border border-base-300 rounded-lg overflow-hidden">
                   <Editor
                     height="400px"
                     defaultLanguage="python"
                     value={createFactorForm.formula}
-                    onChange={(value) => setCreateFactorForm({
-                      ...createFactorForm,
-                      formula: value || ''
-                    })}
+                    onChange={(value) =>
+                      setCreateFactorForm({
+                        ...createFactorForm,
+                        formula: value || '',
+                      })
+                    }
                     options={{
                       readOnly: false,
                       minimap: { enabled: false },
@@ -494,10 +585,18 @@ ${dataAccess}
                   <div>
                     <h6 className="font-medium mb-2">常用函数:</h6>
                     <ul className="space-y-1 text-base-content/70">
-                      <li><code>.pct_change()</code> - 计算收益率</li>
-                      <li><code>.rolling(n).mean()</code> - n日移动平均</li>
-                      <li><code>.rolling(n).std()</code> - n日标准差</li>
-                      <li><code>.shift(n)</code> - 向前/后移动n期</li>
+                      <li>
+                        <code>.pct_change()</code> - 计算收益率
+                      </li>
+                      <li>
+                        <code>.rolling(n).mean()</code> - n日移动平均
+                      </li>
+                      <li>
+                        <code>.rolling(n).std()</code> - n日标准差
+                      </li>
+                      <li>
+                        <code>.shift(n)</code> - 向前/后移动n期
+                      </li>
                     </ul>
                   </div>
                   <div>
@@ -531,7 +630,10 @@ ${dataAccess}
           )}
         </div>
       </div>
-    </div>
+      <form method="dialog" className="modal-backdrop">
+        <button>close</button>
+      </form>
+    </dialog>
   );
 };
